refactor(cli): one module per search verb, and a census that follows
`src/cli/search.ts` was 1453 lines holding every `o2b search` verb -
parsing, validation, execution and rendering, inline, at the highest
cognitive-complexity density in the repository. The verbs now sit under
`src/cli/search/verbs/` with their shared plumbing beside them, which is
the layout `src/cli/brain/` already established rather than a second one
invented here. The dispatcher keeps its path, its single export and its
exit-code mapping.

CLI output is a public contract, so equivalence was measured rather than
argued: 92 invocations were run through the HEAD tree and the new one on
identical fresh vaults, diffing stdout, stderr and exit code byte for
byte across every verb, both streams, empty and populated vaults and
every error path. Zero differences after normalising wall-clock only.

Six duplications removed, including the vault/config/db flag schema
written out fourteen times and the whole JSON-plus-rail emission block
copy-pasted between `index` and `reindex`, comment included.

The terminal-state census had to grow with the split rather than be
excused from it. It walked each handler's call graph within one module,
which was right when a verb was one file and stale once a verb's refusal
could live in a sibling; the walk now follows relative imports inside
the verb's own command subtree, two module hops, explicit. Three
pre-existing hand-classified entries turned out to be detectable under
the wider walk and were removed, leaving three.

The bound is measured, not guessed: following every local import inside
`src/cli` takes detection from 166 of 174 states to all 174, because
every verb parses flags and the flag parser throws - a detector that
cannot fail. The command-subtree limit is what keeps the class meaning
anything. A standing test pins the exact set of states that only the
walk classifies, so a broken import resolution empties that list and
names what would have been silently excused, and the census was proved
non-vacuous by removing a real refusal and watching it fail.
